Add an option to support multiple projects
In the plugins settings there's now a possibility to select multiple projects.
Additionally there are two new options to hide the project selection list and
the room selection list from the meeting_room calendar page.

What's not perfect here is that the rooms are the same for all projects.

There's also a new route added to routes.rb so that
http://[...]/meeting_room_calendar/<project_id>
links directly to the project with id <project_id>.
E.g. http://[..]/meeting_room_calendar/2 will directly fill the calendar
with issues from project '2'.